# openapi.yaml
openapi: 3.0.3
info:
  title: SOJPE C2H Phase 1 API
  version: 1.1.0
  description: |
    Backend for the SOJPE C2H weekly reporting dashboard.
    Runs on AWS Lambda (sojpe-data-pipeline) behind API Gateway; storage in S3 (trigent-c2h-files)
    and DynamoDB (sojpe-reports). The Vercel frontend reaches it via a /api/* rewrite,
    so all paths work both on the gateway host and relative to https://metry360.arkos.studio.

    Reporting periods are derived from the DATA inside uploaded files (the "Period:" header
    in the Selects/Renege reports), not from the upload timestamp.
servers:
  - url: https://2g852sgu1i.execute-api.ap-south-1.amazonaws.com/prod
  - url: https://metry360.arkos.studio

paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: healthy }
                  service: { type: string, example: sojpe-c2h-api }
                  version: { type: string }
                  infrastructure:
                    type: object
                    properties:
                      s3_bucket: { type: string }
                      dynamodb_table: { type: string }
                      region: { type: string }

  /api/upload:
    post:
      summary: Upload the 7 weekly raw Excel files
      description: |
        Files are sent as base64 in a JSON body (not multipart). Each file is stored in S3 under
        reports/{report_id}/{name}. The data period is auto-detected from file contents and
        stored as data_week_ending. The pipeline then runs and the report lands in IN_REVIEW.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [files]
              properties:
                reportId:
                  type: string
                  description: Optional client-generated id (must start with REPORT-)
                files:
                  type: array
                  items:
                    type: object
                    required: [name, content]
                    properties:
                      name: { type: string, example: Weekly Selects Report.xlsx }
                      fileType: { type: string, example: Weekly Selects Report }
                      size: { type: integer }
                      content: { type: string, format: byte, description: base64 file contents }
      responses:
        '202':
          description: Files stored, pipeline started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  report_id: { type: string }
                  status: { type: string, enum: [IN_REVIEW, ERROR] }
                  file_count: { type: integer }
                  data_week_ending: { type: string, nullable: true, example: 29-May-2026 }
                  data_period: { type: string, nullable: true, example: 25-May-2026 to 29-May-2026 }
        '400':
          $ref: '#/components/responses/Error'

  /api/reports:
    get:
      summary: List reports (newest first)
      responses:
        '200':
          description: Reports with metadata and stored file lists
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  count: { type: integer }
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/Report'

  /api/reports/{reportId}:
    get:
      summary: Report details + S3 file listing
      parameters:
        - { name: reportId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Report details }
        '404':
          $ref: '#/components/responses/Error'

  /api/reports/{reportId}/approve:
    post:
      summary: Approve a report (unlocks downloads)
      parameters:
        - { name: reportId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reviewerName: { type: string }
                decision: { type: string }
                notes: { type: string }
      responses:
        '200': { description: Approved }

  /api/reports/{reportId}/reprocess:
    post:
      summary: Re-run the pipeline for a report
      parameters:
        - { name: reportId, in: path, required: true, schema: { type: string } }
      responses:
        '202': { description: Reprocessing started }

  /api/reporting-period:
    get:
      summary: Current reporting week derived from the latest report's data period
      responses:
        '200':
          description: Week/month info driving the dashboard's week-ending labels
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  reporting_period:
                    type: object
                    properties:
                      month: { type: string, example: May 2026 }
                      month_number: { type: integer }
                      week: { type: integer, description: 1-based index into week_endings }
                      week_ending: { type: string, example: 29-May-2026 }
                      week_endings:
                        type: array
                        description: All Fridays of the data month
                        items: { type: string }
                      week_start: { type: string }
                      week_end: { type: string }
                      source:
                        type: string
                        enum: [data_period, uploads, current_date]
                        description: data_period = from file contents; uploads = upload timestamp fallback; current_date = empty database
                      report_id: { type: string, nullable: true }

components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
  schemas:
    Report:
      type: object
      properties:
        report_id: { type: string }
        status: { type: string, enum: [PROCESSING, IN_REVIEW, APPROVED, ERROR] }
        created_at: { type: string, description: UTC ISO timestamp (upload time) }
        month: { type: string, example: May 2026 }
        data_week_ending: { type: string, nullable: true, example: 29-May-2026 }
        data_period: { type: string, nullable: true }
        file_count: { type: integer }
        files:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              size: { type: integer }
              key: { type: string, description: S3 object key }
              file_type: { type: string }